Create NumPy Facet
