+++ pint/facets/numpy/registry.py
@@ -1,0 +1,17 @@
+"""
+    pint.facets.numpy.registry
+    ~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+    :copyright: 2022 by Pint Authors, see AUTHORS for more details.
+    :license: BSD, see LICENSE for more details.
+"""
+
+
+from .quantity import NumpyQuantity
+from .unit import NumpyUnit
+
+
+class NumpyRegistry:
+
+    _quantity_class = NumpyQuantity
+    _unit_class = NumpyUnit
